resource/api: move RestoreError to resource/api
