+++ apiserver/common/errors.go
@@ -132,16 +132,6 @@ func singletonCode(err error) (string, bool) {
 	return code, ok
 }
 
-func singletonError(err error) (error, bool) {
-	errCode := params.ErrCode(err)
-	for singleton, code := range singletonErrorCodes {
-		if errCode == code && singleton.Error() == err.Error() {
-			return singleton, true
-		}
-	}
-	return nil, false
-}
-
 // ServerErrorAndStatus is like ServerError but also
 // returns an HTTP status code appropriate for using
 // in a response holding the given error.
@@ -242,67 +232,4 @@ func DestroyErr(desc string, ids, errs []string) error {
 	}
 	msg = fmt.Sprintf(msg, desc)
 	return errors.Errorf("%s: %s", msg, strings.Join(errs, "; "))
-}
-
-// RestoreError makes a best effort at converting the given error
-// back into an error originally converted by ServerError().
-func RestoreError(err error) error {
-	err = errors.Cause(err)
-
-	if apiErr, ok := err.(*params.Error); !ok {
-		return err
-	} else if apiErr == nil {
-		return nil
-	}
-	if params.ErrCode(err) == "" {
-		return err
-	}
-	msg := err.Error()
-
-	if singleton, ok := singletonError(err); ok {
-		return singleton
-	}
-
-	// TODO(ericsnow) Support the other error types handled by ServerError().
-	switch {
-	case params.IsCodeUnauthorized(err):
-		return errors.NewUnauthorized(nil, msg)
-	case params.IsCodeNotFound(err):
-		// TODO(ericsnow) UnknownModelError should be handled here too.
-		// ...by parsing msg?
-		return errors.NewNotFound(nil, msg)
-	case params.IsCodeAlreadyExists(err):
-		return errors.NewAlreadyExists(nil, msg)
-	case params.IsCodeNotAssigned(err):
-		return errors.NewNotAssigned(nil, msg)
-	case params.IsCodeHasAssignedUnits(err):
-		// TODO(ericsnow) Handle state.HasAssignedUnitsError here.
-		// ...by parsing msg?
-		return err
-	case params.IsCodeNoAddressSet(err):
-		// TODO(ericsnow) Handle isNoAddressSetError here.
-		// ...by parsing msg?
-		return err
-	case params.IsCodeNotProvisioned(err):
-		return errors.NewNotProvisioned(nil, msg)
-	case params.IsCodeUpgradeInProgress(err):
-		// TODO(ericsnow) Handle state.UpgradeInProgressError here.
-		// ...by parsing msg?
-		return err
-	case params.IsCodeMachineHasAttachedStorage(err):
-		// TODO(ericsnow) Handle state.HasAttachmentsError here.
-		// ...by parsing msg?
-		return err
-	case params.IsCodeNotSupported(err):
-		return errors.NewNotSupported(nil, msg)
-	case params.IsBadRequest(err):
-		return errors.NewBadRequest(nil, msg)
-	case params.IsMethodNotAllowed(err):
-		return errors.NewMethodNotAllowed(nil, msg)
-	case params.ErrCode(err) == params.CodeDischargeRequired:
-		// TODO(ericsnow) Handle DischargeRequiredError here.
-		return err
-	default:
-		return err
-	}
 }
