+++ resource/api/client/client.go
@@ -12,7 +12,6 @@ import (
 	"github.com/juju/loggo"
 	charmresource "gopkg.in/juju/charm.v6-unstable/resource"
 
-	"github.com/juju/juju/apiserver/common"
 	"github.com/juju/juju/resource"
 	"github.com/juju/juju/resource/api"
 )
@@ -116,7 +115,7 @@ func (c Client) AddPendingResources(serviceID string, resources []charmresource.
 		return nil, errors.Trace(err)
 	}
 	if result.Error != nil {
-		err := common.RestoreError(result.Error)
+		err := api.RestoreError(result.Error)
 		return nil, errors.Trace(err)
 	}
